requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0
tenacity==8.2.3
aiobreaker==1.2.0